# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A personal geek media tracker API - Track anime, movies, series, games, manga, books and music\nFeatures: Personal lists, progress tracking, ratings, favorites, and more",
        "title": "Geekery API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Rafael C Ribeiro",
            "email": "support@geekery.dev"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/items": {
            "get": {
                "description": "Get paginated list of all items from the global catalog",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Get all items",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "anime",
                            "movie",
                            "series",
                            "game",
                            "manga",
                            "light_novel",
                            "music",
                            "book"
                        ],
                        "type": "string",
                        "description": "Filter by media type",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success - returns paginated items",
                        "schema": {
                            "$ref": "#/definitions/github_com_rafaelc-rb_geekery-api_internal_dto.PaginatedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new item in the global catalog (admin only in the future)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Create item",
                "parameters": [
                    {
                        "description": "Item to create",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_rafaelc-rb_geekery-api_internal_models.Item"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Item created successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_rafaelc-rb_geekery-api_internal_models.Item"
                        }
                    },
                    "400": {
                        "description": "Bad request - validation error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/items/search": {
            "get": {
                "description": "Search items by title (case-insensitive) with pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Search items",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success - returns matching items",
                        "schema": {
                            "$ref": "#/definitions/github_com_rafaelc-rb_geekery-api_internal_dto.PaginatedResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/items/{id}": {
            "get": {
                "description": "Get a specific item from the catalog by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Get item by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success - returns item",
                        "schema": {
                            "$ref": "#/definitions/github_com_rafaelc-rb_geekery-api_internal_models.Item"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Item not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing item in the catalog (admin only in the future)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Update item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Item data to update",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_rafaelc-rb_geekery-api_internal_models.Item"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item updated successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - validation error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Item not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an item from the catalog (admin only in the future)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Delete item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Item deleted successfully"
                    },
                    "400": {
                        "description": "Bad request - invalid ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Item not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_rafaelc-rb_geekery-api_internal_dto.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "pagination": {
                    "$ref": "#/definitions/github_com_rafaelc-rb_geekery-api_internal_dto.PaginationMeta"
                }
            }
        },
        "github_com_rafaelc-rb_geekery-api_internal_dto.PaginationMeta": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "has_next": {
                    "type": "boolean"
                },
                "has_previous": {
                    "type": "boolean"
                },
                "items_per_page": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "github_com_rafaelc-rb_geekery-api_internal_models.AnimeData": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "episodes": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "item_id": {
                    "description": "FK para items",
                    "type": "integer"
                },
                "studio": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_rafaelc-rb_geekery-api_internal_models.BookData": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "chapters": {
                    "description": "Total de capítulos",
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "item_id": {
                    "description": "FK para items",
                    "type": "integer"
                },
                "pages": {
                    "description": "Total de páginas",
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "volumes": {
                    "description": "Total de volumes (manga/LN sempre tem, books às vezes)",
                    "type": "integer"
                }
            }
        },
        "github_com_rafaelc-rb_geekery-api_internal_models.GameData": {
            "type": "object",
            "properties": {
                "average_playtime": {
                    "description": "Horas médias para completar",
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "developer": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "item_id": {
                    "description": "FK para items",
                    "type": "integer"
                },
                "platform": {
                    "description": "PC, PS5, Switch, Xbox, etc",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_rafaelc-rb_geekery-api_internal_models.Item": {
            "type": "object",
            "properties": {
                "anime_data": {
                    "description": "Dados específicos por tipo (apenas um será não-nil baseado no Type)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_rafaelc-rb_geekery-api_internal_models.AnimeData"
                        }
                    ]
                },
                "book_data": {
                    "$ref": "#/definitions/github_com_rafaelc-rb_geekery-api_internal_models.BookData"
                },
                "cover_url": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "external_metadata": {
                    "description": "Metadados de APIs externas (MAL, IMDb, etc)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_rafaelc-rb_geekery-api_internal_models.JSONB"
                        }
                    ]
                },
                "game_data": {
                    "$ref": "#/definitions/github_com_rafaelc-rb_geekery-api_internal_models.GameData"
                },
                "id": {
                    "type": "integer"
                },
                "movie_data": {
                    "$ref": "#/definitions/github_com_rafaelc-rb_geekery-api_internal_models.MovieData"
                },
                "music_data": {
                    "$ref": "#/definitions/github_com_rafaelc-rb_geekery-api_internal_models.MusicData"
                },
                "release_date": {
                    "description": "Data de lançamento/estreia",
                    "type": "string"
                },
                "series_data": {
                    "$ref": "#/definitions/github_com_rafaelc-rb_geekery-api_internal_models.SeriesData"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_rafaelc-rb_geekery-api_internal_models.Tag"
                    }
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/github_com_rafaelc-rb_geekery-api_internal_models.MediaType"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_rafaelc-rb_geekery-api_internal_models.JSONB": {
            "type": "object",
            "additionalProperties": true
        },
...
        "github_com_rafaelc-rb_geekery-api_internal_models.MediaType": {
            "type": "string",
            "enum": [
                "anime",
                "movie",
                "series",
                "game",
                "manga",
                "light_novel",
                "music",
                "book"
            ],
            "x-enum-varnames": [
                "MediaTypeAnime",
                "MediaTypeMovie",
                "MediaTypeSeries",
                "MediaTypeGame",
                "MediaTypeManga",
                "MediaTypeLightNovel",
                "MediaTypeMusic",
                "MediaTypeBook"
            ]
        },
        "github_com_rafaelc-rb_geekery-api_internal_models.MovieData": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "director": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "item_id": {
                    "description": "FK para items",
                    "type": "integer"
                },
                "runtime": {
                    "description": "em minutos",
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_rafaelc-rb_geekery-api_internal_models.MusicData": {
            "type": "object",
            "properties": {
                "album": {
                    "type": "string"
                },
                "artist": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "duration": {
                    "description": "Duração total do álbum em segundos",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "item_id": {
                    "description": "FK para items",
                    "type": "integer"
                },
                "tracks": {
                    "description": "Número de faixas",
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "github_com_rafaelc-rb_geekery-api_internal_models.SeriesData": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "episodes": {
                    "description": "Total de episódios",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "item_id": {
                    "description": "FK para items",
                    "type": "integer"
                },
                "seasons": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_rafaelc-rb_geekery-api_internal_models.Tag": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_rafaelc-rb_geekery-api_internal_models.Item"
                    }
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token (future implementation)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}